Sửa lỗi hiện product option với sản phẩm Brudo

+++ LibMagestore/src/main/java/com/magestore/app/pos/model/catalog/PosProductOptionPriceConfigPriceAmount.java
@@ -5,6 +5,6 @@ package com.magestore.app.pos.model.catalog;
  */
 
 public class PosProductOptionPriceConfigPriceAmount {
-    int amount;
+    float amount;
     Object[] adjustments;
 }
